refactor(api): refactor stats endpoints
