+++ apps/cron-job-manager/src/http/create-http-agents.ts
@@ -1,0 +1,20 @@
+import Agent from "agentkeepalive";
+
+// TODO: Configure these values for localdev and production envs
+export function createHttpAgent() {
+  return new Agent({
+    maxSockets: 100,
+    maxFreeSockets: 10,
+    timeout: 60000,
+    freeSocketTimeout: 30000,
+  });
+}
+
+export function createHttpsAgent() {
+  return new Agent.HttpsAgent({
+    maxSockets: 100,
+    maxFreeSockets: 10,
+    timeout: 60000,
+    freeSocketTimeout: 30000,
+  });
+}
