+++ packages/api/src/utils/time.ts
@@ -1,0 +1,3 @@
+export function toTimestamp(date: Date): number {
+  return Math.floor(date.getTime() / 1000);
+}
